Initial port
This code was originall developed as a Haskell implementation of Fagin,
a large bioinformatics pipeline. Here I have copied over the code, but
the commit history remains in the Fagin repo (which is currently
private).

There is a lot of valuable stuff in here. I will probably repurpose it
eventually.